+++ crates/miden-lib/src/errors/transaction_errors.rs
@@ -67,6 +67,15 @@ pub enum TransactionKernelError {
     AccountStorageSlotsNumMissing(u32),
     #[error("account nonce can only be incremented once")]
     NonceCanOnlyIncrementOnce,
+    /// This variant signals that a signature over the contained commitments is required, but
+    /// missing.
+    #[error("transaction requires a signature")]
+    Unauthorized {
+        account_delta_commitment: Word,
+        input_notes_commitment: Word,
+        output_notes_commitment: Word,
+        salt: Word,
+    },
 }
 
 // TRANSACTION EVENT PARSING ERROR