+++ crates/miden-objects/src/transaction/mod.rs
@@ -15,6 +15,7 @@ mod proven_tx;
 mod transaction_id;
 mod tx_args;
 mod tx_header;
+mod tx_summary;
 mod tx_witness;
 
 pub use executed_tx::{ExecutedTransaction, TransactionMeasurements};
@@ -28,4 +29,5 @@ pub use proven_tx::{
 pub use transaction_id::TransactionId;
 pub use tx_args::{TransactionArgs, TransactionScript};
 pub use tx_header::TransactionHeader;
+pub use tx_summary::TransactionSummary;
 pub use tx_witness::TransactionWitness;