+++ crates/miden-tx/src/executor/exec_host.rs
@@ -76,6 +76,11 @@ impl<'store, 'auth> TransactionExecutorHost<'store, 'auth> {
     // PUBLIC ACCESSORS
     // --------------------------------------------------------------------------------------------
 
+    /// Returns a reference to the underlying [`TransactionBaseHost`].
+    pub(super) fn base_host(&self) -> &TransactionBaseHost {
+        &self.base_host
+    }
+
     /// Returns a reference to the `tx_progress` field of this transaction host.
     pub fn tx_progress(&self) -> &TransactionProgress {
         self.base_host.tx_progress()